<fix>[zstack]: fix incorrect zstack code

See merge request zstackio/zstack!3865

+++ core/src/main/java/org/zstack/core/tacker/PingTracker.java
@@ -163,7 +163,7 @@ public abstract class PingTracker implements Component {
         if (CoreGlobalProperty.UNIT_TEST_ON) {
             trackerThread = thdf.submitPeriodicTask(new Tracker(), getPingInterval());
         } else {
-            trackerThread = thdf.submitPeriodicTask(new Tracker(), getPingInterval() + new Random().nextInt(30));
+            trackerThread = thdf.submitPeriodicTask(new Tracker(), (long) getPingInterval() + new Random().nextInt(30));
         }
     }
